\name{fractal-art-package}
\alias{fractal-art-package}
\alias{fractal-art}
\docType{package}
\title{
\packageTitle{fractal-art}
}
\description{
\packageDescription{fractal-art}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fractal-art}
\packageIndices{fractal-art}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fractal-art}

Maintainer: \packageMaintainer{fractal-art}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
